+++ components/MenuDetails.js
@@ -8,7 +8,20 @@
 //c) visualizza i dettagli del menu selezionato
 //d) visualizza un pulsante per acquistare il menu
 //e) gestisce la pressione del pulsante di acquisto chiamando la funzione compraMenu
-
+/*COME FUNZIONA LA LOGICA DI ACQUISTO DEL MENU:
+- Nella componente MenuDetails appare il pulsante "compra menu",
+- Quando il pulsante "compra menu" viene premuto appare un alert con una richiesta di conferma, con i relativi pulsanti "conferma" o "annulla"
+- Se viene premuto il pulsante "annulla", l'alert scompare e si torna alla pagina MenuDetails
+- Se viene premuto il pulsante "conferma", viene invocata la funzione "compraMenu" definita a livello di ViewModel, chiamata a livello di App.js e passata a MenuDetails come props.
+- La funzione compraMenu() esegue le seguenti azioni:
+    a) controlla che l'utente sia registrato tramite il controllo (già previsto) isRegistered, se l'utente non è registrato porta alla componente "UserProfile"
+    b) controlla che l'utente non abbia ordini in corso facendo una chiamata a server definita in ApiService.js: isOrderActive(), 
+    se c'è un'ordine in corso appare un alert che segnala che c'è un ordine in corso e non è possibile piazzarne un altro
+    c) se l'utente è registrato e non ci sono ordini in corso, effettua un ordine tramite una chiamata a server definita in ApiService.js: makeOrder()
+- TODO: Lavorare su ApiService per scrivere le funzioni necessarie per fare un ordine:
+    a) makeOrder()
+    b) isOrderActive()
+*/
 import React, { useState, useEffect } from 'react';
 import { View, Text, Button, StyleSheet, Image } from 'react-native';
 import ViewModel from '../ViewModel';
@@ -16,7 +29,7 @@ import AppStyles from '../AppStyles';
 import { ScrollView } from 'react-native';
 
 
-export default function MenuDetails({menu, sid, onBack, userLocation}) {
+export default function MenuDetails({menu, sid, onBack, userLocation, onOrder}) {
 
     const [detailedMenu, setDetailedMenu] = useState(null);
     const [loading, setLoading] = useState(true);
@@ -50,14 +63,16 @@ export default function MenuDetails({menu, sid, onBack, userLocation}) {
     console.log("Menu da visualizzare: {nome:", detailedMenu.name, ", prezzo: ", detailedMenu.price, ", tempo: ", detailedMenu.deliveryTime, ", descrizione: ", detailedMenu.longDescription);
     return (
         <View style={AppStyles.menuCard}>
-            <Button title="Torna alla lista" onPress={onBack} color="#FF7F00" />
             <Text style={AppStyles.menuCardTitle} >{detailedMenu.name}</Text>
-            <Text style={AppStyles.menuDetailsDescription} >{detailedMenu.longDescription}</Text>
+            <Text style={AppStyles.menuCardDescription} >{detailedMenu.longDescription}</Text>
             <Text style={AppStyles.menuCardPrice} >€{detailedMenu.price}</Text>
             <Text style={AppStyles.menuCardDeliveryTime} >
                 Tempo di consegna: {detailedMenu.deliveryTime} min
             </Text>
-            <Image source={{ uri: detailedMenu.image }} style={AppStyles.menuDetailsImage} />
+            <Image source={{ uri: detailedMenu.image }} style={AppStyles.menuCardImage} />
+            <Button title="Compra menu" onPress={() => onOrder(menu.mid, userLocation)} color="blue" />
+            <Button title="Torna alla lista" onPress={onBack} color="#FF7F00" />
+
         </View>
       );
