+++ components/MenuList.js
@@ -4,11 +4,11 @@ import React from 'react';
 import { ScrollView } from 'react-native';
 import MenuCard from './MenuCard';
 
-export default function MenuList({ menus, onSeeDetails }) {
+export default function MenuList({ menus, onSeeDetails}) {
   return (
     <ScrollView contentContainerStyle={AppStyles.menuListContainer}>
         {menus.map((menu) => (
-            <MenuCard key={menu.mid} menu={menu} onSeeDetails={onSeeDetails} />
+            <MenuCard key={menu.mid} menu={menu} onSeeDetails={onSeeDetails}  />
         ))}
     </ScrollView>
   );
