+++ models/objects/DetailedMenu.js
