--- src/models/objects/DetailedMenu.js
@@ -1,9 +1,0 @@
-// DetailedMenu.js
-import Menu from './Menu';
-
-export default class DetailedMenu extends Menu {
-    constructor(mid, name, shortDescription, price, deliveryTime, imageVersion, location, longDescription) {
-        super(mid, name, shortDescription, price, deliveryTime, imageVersion, location);
-        this.longDescription = longDescription;
-    }
-}
